fix: up to case-data 10 and moved mocked-data as the build broke
