+++ backend/src/config/api-config.ts
@@ -11,7 +11,7 @@ export const APIS = [
   // },
   {
     name: 'case-data',
-    version: '9.0',
+    version: '10.0',
   },
   {
     name: 'casestatus',
